Update versioned_docs/version-3.0.6/todo-list.md
Fix "If you do not provide children, you must specify an aria-label for accessibility" error. TBH, I don't know if I put the correct value to it, because based on my experience on a web based app, Label should be labeled after the input name. And FYI, this fix works on my end